fix: tag non-stable settings with new system (#24287)
This PR changes some setting tags as follows:

- Remove the experimental tag from the `python.missingPackage.severity`
setting considering the setting does not seem to be experimental and
that the setting does not seem to be part of any experiment.
- Add the onExP tag onto the `python.locator` setting considering the
setting is both experimental (based on its description) and part of an
experiment.
- Change the experimental tag to preview for
`python.terminal.shellIntegration.enabled`, considering it is a
non-stable setting that does not seem to be part of an experiment.
- Add the onExP tag to `python.REPL.sendToNativeREPL` considering it is
part of an experiment.